[S06P12E204-226] 미팅신청-BE : 스타가 신청한 팬미팅을 수정할 때 사용 하는 API 구현
스타가 신청한 팬미팅을 수정할 때 사용
수정을 요청한 팬미팅이 존재 하지 않으면 오류를 리턴

+++ Backend/src/main/java/com/ssafy/yourstar/domain/meeting/db/repository/ApplicantRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplicantRepository extends JpaRepository<Applicant, Long> {
+public interface ApplicantRepository extends JpaRepository<Applicant, Integer> {
 }
